Fix span emplace_back size type mismatch for C++20 std::span compatibility

+++ include/dds.hpp
@@ -578,7 +578,7 @@ namespace dds {
                 uint32_t size = computeMipmapSize(image->format, width, height);
                 totalSize += static_cast<uint64_t>(size);
 
-                image->mipmaps.emplace_back(ptr, size);
+                image->mipmaps.emplace_back(ptr, static_cast<std::size_t>(size));
                 ptr += size;
 
                 width = max(width / 2, 1u);
